Makes the BuildMeshVertexDataSystem an image based system.
