+++ Core/Systems/Render/BuildImageVertexDataSystem.cs.meta
