+++ Core/Systems/Render/MeshCacheSystem.cs
@@ -9,7 +9,7 @@ namespace UGUIDots.Render.Systems {
     // I'm not entirely convinced this achitecture would work, it does not include batching of meshes which would be
     // important to reduce the # of issued draw calls to the GPU...
     [UpdateInGroup(typeof(MeshBatchingGroup))]
-    [UpdateAfter(typeof(BuildMeshVertexDataSystem))]
+    [UpdateAfter(typeof(BuildImageVertexDataSystem))]
     public class MeshCacheSystem : ComponentSystem {
 
         public struct MeshPropertyPair : IEquatable<MeshPropertyPair> {
